Update: move `algorithms` to the root
