+++ cmaze/src/gameboard/mod.rs
@@ -2,6 +2,3 @@ pub mod maze;
 pub use maze::Maze;
 pub mod cell;
 pub use cell::{Cell, CellWall};
-
-// move to root
-pub mod algorithms;
