+++ cmaze/src/lib.rs
@@ -3,3 +3,4 @@ pub mod gameboard;
 pub mod dims;
 pub mod progress;
 pub mod array;
+pub mod algorithms;
